Add files via upload
The short-term approach to forecasting water demand included regression models, time series analysis, and the Artificial Neural Network (ANN) technique. These methods were used to evaluate the daily frequency data of rainfall, maximum air temperature, and the one-period lags for Wrexham, Ireland. The model was split into a two-third training dataset, to cover the data the model knows and a third remnant for the validation.